pyproject.toml
[project]
name = "cockpit-crm-core-django"
version = "0.1.0"
description = ""
authors = [
    {name = "ketstap162",email = "zp162162@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)"
]

[tool.black]
line-length = 88
target-version = ['py312']
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''
